doc: restore documentation build
Add doc/Makefile for installing Sphinx into a virtualenv and doing a
build.  We probably don't want to commit the built docs to the repo
(they're 5.8 MB, mostly fonts), but we also weren't doing that before.
Without committing the files, though, `ninja dist` won't include the
generated docs in the source tarball, which is a change from Automake.