meson: skip dist script if we're a subproject
Dist scripts can't be used in subprojects before Meson 0.58.0.  In our
case, it's not clear that it makes sense to do so.  The parent project
probably doesn't want to ship our documentation, may not have vendored
our dotfiles, and may not want us skipping files they did choose to
vendor.