doc: switch back to libclang to avoid hardcoding LLVM version
The libclang package has an embedded copy of libclang, which allows us to
avoid requiring the system to have a particular clang version.  This
didn't work before because the lib_search_dirs logic in conf.py was
clobbering a working path with a broken one.  That logic is no longer
needed, so just remove it.